+++ src/xtd.forms/src/xtd/forms/combo_box.cpp
@@ -22,14 +22,14 @@ combo_box::combo_box() {
   drop_down_height_ = static_cast<int32_t>(font().get_height()) * 9;
 
   items_.item_added += [&](size_t pos, const item& item) {
-    native::combo_box::insert_item(handle(), pos, item.value());
+    if (is_handle_created()) native::combo_box::insert_item(handle(), pos, item.value());
     combo_box::item selected_item;
     if (selected_index_ != npos && selected_index_ < items_.size()) selected_item = items_[selected_index_];
     this->selected_item(selected_item);
   };
 
   items_.item_removed += [&](size_t pos, const item& item) {
-    native::combo_box::delete_item(handle(), pos);
+    if (is_handle_created()) native::combo_box::delete_item(handle(), pos);
 
     combo_box::item selected_item;
     if (selected_index_ != npos && selected_index_ < items_.size()) selected_item = items_[selected_index_];
@@ -37,7 +37,7 @@ combo_box::combo_box() {
   };
   
   items_.item_updated += [&](size_t pos, const item& item) {
-    native::combo_box::update_item(handle(), pos, item.value());
+    if (is_handle_created()) native::combo_box::update_item(handle(), pos, item.value());
     combo_box::item selected_item;
     if (selected_index_ != npos && selected_index_ < items_.size()) selected_item = items_[selected_index_];
     this->selected_item(selected_item);
@@ -57,7 +57,7 @@ list_control& combo_box::selected_index(size_t selected_index) {
   if (selected_index_ != selected_index) {
     if (selected_index != npos && selected_index >= items_.size()) throw argument_out_of_range_exception("Selected index greater than items size"_t, current_stack_frame_);
     selected_index_ = selected_index;
-    native::combo_box::selected_index(handle(), selected_index_);
+    if (is_handle_created()) native::combo_box::selected_index(handle(), selected_index_);
     
     item selected_item;
     if (selected_index_ != npos) selected_item = items_[selected_index_];
@@ -92,11 +92,11 @@ combo_box& combo_box::sorted(bool sorted) {
 }
 
 void combo_box::begin_update() {
-  native::combo_box::begin_update(handle());
+  if (is_handle_created()) native::combo_box::begin_update(handle());
 }
 
 void combo_box::end_update() {
-  native::combo_box::end_update(handle());
+  if (is_handle_created()) native::combo_box::end_update(handle());
 }
 
 forms::create_params combo_box::create_params() const {
